requirements.txt
alabaster==0.7.12
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.2
asgiref==3.7.2
astroid==2.11.2
asttokens==2.0.5
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
bcrypt==4.0.1
binaryornot==0.4.4
black==22.3.0
blinker==1.6.3
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.3.1
colorama==0.4.4
comm==0.2.2
contourpy==1.2.0
cookiecutter==1.7.3
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
crispy-bootstrap5==0.6
cryptography==36.0.2
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.4
distlib==0.3.4
Django==5.0.1
django-allauth==0.50.0
django-cors-headers==3.11.0
django-coverage-plugin==2.0.2
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-extensions==3.1.5
django-model-utils==4.2.0
django-redis==5.2.0
django-stubs==1.9.0
django-stubs-ext==0.4.0
djangorestframework==3.13.1
djangorestframework-stubs==1.4.0
dnspython==2.8.0
docutils==0.17.1
drf-spectacular==0.22.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
executing==0.8.3
factory-boy==3.2.1
Faker==13.3.4
fastapi==0.128.0
filelock==3.6.0
flake8==4.0.1
flake8-isort==4.1.1
Flask==3.0.0
fonttools==4.50.0
greenlet==3.3.0
h11==0.16.0
httptools==0.6.4
identify==2.4.12
idna==3.11
imagesize==1.3.0
inflection==0.5.1
iniconfig==1.1.1
ipdb==0.13.9
ipykernel==6.29.4
ipython==8.2.0
isort==5.10.1
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jinja2-time==0.2.0
jsonschema==4.4.0
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy-object-proxy==1.7.1
livereload==2.6.3
MarkupSafe==2.1.1
matplotlib==3.8.3
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy==0.942
mypy-extensions==0.4.3
nest-asyncio==1.6.0
nodeenv==1.6.0
numpy==1.26.4
oauthlib==3.2.0
packaging==21.3
pandas==2.2.1
pandasql==0.7.3
parso==0.8.3
passlib==1.7.4
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.1
pluggy==1.0.0
poyo==0.5.0
pre-commit==2.18.1
prompt-toolkit==3.0.28
psutil==5.9.8
psycopg2==2.9.3
psycopg2-binary==2.9.11
pure-eval==0.2.2
py==1.11.0
pyasn1==0.6.2
pycodestyle==2.8.0
pycparser==2.21
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyflakes==2.4.0
Pygments==2.11.2
PyJWT==2.3.0
pylint==2.13.4
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==7.1.1
pytest-django==4.5.2
pytest-sugar==0.9.4
python-dateutil==2.8.2
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.27
python-slugify==6.1.1
python3-openid==3.2.0
pytz==2022.1
pywin32==306
PyYAML==6.0
pyzmq==25.1.2
redis==4.2.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==4.5.0
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==2.0.45
sqlparse==0.4.2
stack-data==0.2.0
starlette==0.50.0
termcolor==1.1.0
testfixtures==6.18.5
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tornado==6.4
traitlets==5.14.2
types-pytz==2021.3.6
types-PyYAML==6.0.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.40.0
virtualenv==20.14.0
virtualenvwrapper-win==1.2.7
watchdog==2.1.7
watchfiles==1.0.5
wcwidth==0.2.5
websockets==15.0.1
Werkzeug==3.0.1
whitenoise==6.0.0
wrapt==1.14.0
